requirements.txt
Flask>=2.0
Flask-Login
Flask-WTF
Flask-Migrate
Flask-SQLAlchemy
python-dotenv
pymysql
Werkzeug
Pillow
gunicorn
